+++ src/table.rs
@@ -804,9 +804,12 @@ pub trait ReadableTable<K: Key + 'static, V: Value + 'static>: ReadableTableMeta
 }
 
 /// A read-only untyped table
+///
+/// Keeps its read transaction alive until the table is dropped.
 pub struct ReadOnlyUntypedTable {
     name: String,
     tree: RawBtree,
+    _transaction_guard: Arc<TransactionGuard>,
 }
 
 impl Sealed for ReadOnlyUntypedTable {}
@@ -841,14 +844,21 @@ impl ReadOnlyUntypedTable {
     pub(crate) fn new(
         name: &str,
         root_page: Option<BtreeHeader>,
-        hint: PageHint,
         fixed_key_size: Option<usize>,
         fixed_value_size: Option<usize>,
+        guard: Arc<TransactionGuard>,
         mem: PageResolver,
     ) -> Self {
         Self {
             name: name.to_string(),
-            tree: RawBtree::new(root_page, fixed_key_size, fixed_value_size, mem, hint),
+            tree: RawBtree::new(
+                root_page,
+                fixed_key_size,
+                fixed_value_size,
+                mem,
+                PageHint::Clean,
+            ),
+            _transaction_guard: guard,
         }
     }
 }